feat:componentes iniciais da aplicação
