+++ src/App.jsx
@@ -1,0 +1,11 @@
+import Tasks from "./components/Tasks.jsx";
+
+const App = () => {
+  return (
+    <>
+      {" "}
+      <Tasks />{" "}
+    </>
+  );
+};
+export default App;
