+++ src/components/Tasks.scss
@@ -1,0 +1,35 @@
+.tasks-container {
+  padding: 60px 85px;
+  display: flex;
+  flex-direction: column;
+  max-width: 40%;
+  padding-top: 50px;
+  height: 100vh;
+
+  h2 {
+    margin-bottom: 24px;
+  }
+
+  h3 {
+    font-size: 18px;
+  }
+
+  .last-tasks {
+    flex: 1;
+    overflow: hidden;
+    display: flex;
+    flex-direction: column;
+    .tasks-lists {
+      margin-top: 20px;
+    }
+  }
+
+  .completed-tasks {
+    flex: 1;
+    padding-top: 28px;
+
+    .tasks-lists {
+      margin-top: 28px;
+    }
+  }
+}
